Support file mounts on Windows 1803 and above
Windows 1803 introduces the bind filter driver, bindflt.sys, which
introduces several mounting improvements over the previous directory
symbolic link approach. These improvements include:

- No longer appear as symbolic links, improving compatibilty.
- Can be mounted on top of existing directories and files.
- Can be files, instead of only directories.

This change addresses the last point by allowing non-directory file
mounts on Windows 1803 (build 17134) and greater. The first two
improvements already work without this, or any other change.

This change also rearranges the code such that the Windows and LCOW
mount parsers are only built for, and tested on Windows, and the Linux
parser is only built for, and tested on Linux. This allows the Windows
parser to call system.GetOSVersion, which only exists for Windows.
