+++ volume/mounts/lcow_parser.go
@@ -1,3 +1,5 @@
+// +build windows
+
 package mounts // import "github.com/docker/docker/volume/mounts"
 
 import (
