+++ volume/mounts/linux_parser.go
@@ -1,3 +1,5 @@
+// +build !windows
+
 package mounts // import "github.com/docker/docker/volume/mounts"
 
 import (
@@ -107,12 +109,6 @@ func (p *linuxParser) validateMountConfigImpl(mnt *mount.Mount, validateBindSour
 		return &errMountConfig{mnt, errors.New("mount type unknown")}
 	}
 	return nil
-}
-
-// read-write modes
-var rwModes = map[string]bool{
-	"rw": true,
-	"ro": true,
 }
 
 // label modes
