+++ volume/mounts/volume_unix.go
@@ -13,6 +13,7 @@ func (p *linuxParser) HasResource(m *MountPoint, absolutePath string) bool {
 	return err == nil && relPath != ".." && !strings.HasPrefix(relPath, fmt.Sprintf("..%c", filepath.Separator))
 }
 
-func (p *windowsParser) HasResource(m *MountPoint, absolutePath string) bool {
-	return false
+// NewParser creates a parser for a given container OS, depending on the current host OS (linux on a windows host will resolve to an lcowParser)
+func NewParser(containerOS string) Parser {
+	return &linuxParser{}
 }
